handle HELP case as a special one
